==> notebooks/data_prep/step_1.ipynb <==
## Preprep \#1

### Load of Data Raw or processed

import numpy as np
data_save = True
import xmltodict

languages = ['en', 'jp','fr']

def train_prep_data(language):
    categories = ['dvd', 'books', 'music']
    data = {}
    for category in categories:
        path_train = f'../../data/amazon_reviews/cls-acl10-unprocessed/{language}/{category}/train.review'
        file_train = open(path_train,"r")
        xml_train= file_train.read()
        ordered_train=xmltodict.parse(xml_train)
        train_ordered_data = ordered_train['items']['item'] # extract dictionary        
        
        
        #unlabeled data
        path_unlabeled = f'../../data/amazon_reviews/cls-acl10-unprocessed/{language}/{category}/unlabeled.review'
        file_unlabeled = open(path_unlabeled,"r")
        xml_unlabeled= file_unlabeled.read()
        ordered_unlabeled=xmltodict.parse(xml_unlabeled)        
        unlabeled_ordered_data = ordered_unlabeled['items']['item'] # extract dictionary
        
        key_in_dict = list(train_ordered_data[0].keys())
        i = len(data)
        for entry in train_ordered_data:
            keys = {j:None for j in key_in_dict}
            data[i] = keys
            for key, value in entry.items():
                data[i][key] = value
            i+=1
            
        for entry in unlabeled_ordered_data:
            keys = {j:None for j in key_in_dict}
            data[i] = keys
            for key, value in entry.items():
                data[i][key] = value
            i+=1
    return data

train_prepped_data = {}
for lang in languages:
    train_prepped_data[lang] = train_prep_data(lang)


def test_prep_data(language):
    categories = ['dvd', 'books', 'music']
    data = {}
    for category in categories:
        path_train = f'../../data/amazon_reviews/cls-acl10-unprocessed/{language}/{category}/unlabeled.review'

        fileptr = open(path_train,"r")
        xml_content= fileptr.read()
        my_ordered_dict=xmltodict.parse(xml_content)
        
        ordered_data = my_ordered_dict['items']['item'] # extract dictionary

        
        key_in_dict = list(ordered_data[0].keys())
        i = len(data)
        for entry in ordered_data:
            keys = {j:None for j in key_in_dict}
            data[i] = keys
            for key, value in entry.items():

                data[i][key] = value
            i+=1
    return data

test_prepped_data = {}
for lang in languages:
    test_prepped_data[lang] = test_prep_data(lang)

if data_save == True:
    import pickle 

    with open(f'../../data/amazon_reviews/train/data_train.json', 'wb') as file:
        pickle.dump(train_prepped_data, file)
    with open(f'../../data/amazon_reviews/test/data_test.json', 'wb') as file:
        pickle.dump(test_prepped_data, file)
else:
    print('Did nothing blud')
